+++ utils.js
@@ -117,7 +117,7 @@ function formatLabel( template, node )
 {
     var result = template;
 
-    result = result.replace( /\$\{line\}/g, node.line );
+    result = result.replace( /\$\{line\}/g, ( node.line + 1 ) );
     result = result.replace( /\$\{column\}/g, node.column );
     result = result.replace( /\$\{tag\}/g, node.tag );
     result = result.replace( /\$\{after\}/g, node.after );